+++ azure-storage-blob/ChangeLog.md
@@ -5,6 +5,7 @@
 ## Version XX.XX.XX:
 - Support for 2018-11-09 REST version. Please see our REST API documentation and blog for information about the related added features.
 - Added support for append block from URL(synchronously) for append blobs.
+- Added support for update page from URL(synchronously) for page blobs.
 - Added support for generating and using blob snapshot SAS tokens.
 - Added support for generating user delegation SAS tokens.